Fix types

+++ src/attention_predictor.rs
@@ -395,7 +395,7 @@ impl Model {
 
             // Add some noise to input
             let input_shape = input.shape().clone();
-            let input = (input + Tensor::randn( 0.0, 0.001, input_shape, device)?)?;
+            let input = (input + Tensor::randn( 0.0 as f32, 0.001 as f32, input_shape, device)?)?;
 
             inputs.push(input);
             targets.push(target.squeeze(0)?);
